# clients/python/archiver-openapi.json
{
  "components": {
    "schemas": {
...
      "EnvelopeResponse": {
        "description": "Outer ``{\"detail\": ErrorEnvelope}`` wrapper for OpenAPI docs.\n\nPublic name (no leading underscore) so ``openapi-python-client`` generates a\ncleanly-named SDK model \u2014 the class name is what surfaces in\n``components/schemas`` and feeds the SDK code generator.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorEnvelope"
          }
        },
        "required": [
          "detail"
        ],
        "title": "EnvelopeResponse",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Unified error response body.",
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional kind-specific structured payload (e.g. conflict id).",
            "title": "Data"
          },
          "errors": {
            "description": "Field-level problems; empty list when none apply.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "kind": {
            "description": "Discriminator for client-side switching.",
            "enum": [
              "body",
              "schema",
              "domain",
              "lookup",
              "conflict",
              "auth",
              "unimplemented",
              "server"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "description": "Human-readable summary; safe to surface to users.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Page_InfoItemOut_": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InfoItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "has_more",
          "limit",
          "offset"
        ],
        "title": "Page[InfoItemOut]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "archiver",
    "version": "4.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/info-items": {
      "get": {
        "description": "List InfoItems with offset pagination, including active sources and rep_spec assignments.\n\n``has_more`` is derived via a ``limit+1`` probe; no total count is computed.\nActive ``info_item_sources`` and ``info_item_rep_specs`` are batch-loaded in\ntwo additional queries (not N+1).",
        "operationId": "list_info_items_api_v1_info_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 9223372036854775807,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InfoItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Bad request (envelope)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Auth required (envelope)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Forbidden (envelope)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Not found (envelope)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Conflict (envelope)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Validation failed (envelope)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Internal error (envelope)."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Info Items",
        "tags": [
          "info-items"
        ]
      },
      "post": {
        "description": "Create an InfoItem.\n\nOptionally accepts ``initial_url`` + ``initial_source_specs`` (atomically creates\na primary InfoSource binding) and ``initial_rep_spec_assignments`` (creates\neffective-dated RepSpec assignments). All writes are a single transaction; any\nvalidation or lookup failure rolls back the whole thing.",
        "operationId": "create_info_item_api_v1_info_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfoItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Bad request (envelope)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Auth required (envelope)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Forbidden (envelope)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Not found (envelope)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Conflict (envelope)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Validation failed (envelope)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Internal error (envelope)."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Info Item",
        "tags": [
          "info-items"
        ]
      }
    },
    "/api/v1/info-items/{info_item_id}": {
      "get": {
        "description": "Fetch a single InfoItem by ID, including sources and active rep_spec assignments.\n\nPass ``include_deactivated=true`` to also include previous primaries and other\ndeactivated source bindings in ``info_item_sources``.",
        "operationId": "get_info_item_api_v1_info_items__info_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "info_item_id",
            "required": true,
            "schema": {
              "title": "Info Item Id",
              "type": "string"
            }
          },
          {
            "description": "When true, include deactivated bindings (previous primaries and other deactivated sources) in info_item_sources. When false (default), only active bindings are returned.",
            "in": "query",
            "name": "include_deactivated",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include deactivated bindings (previous primaries and other deactivated sources) in info_item_sources. When false (default), only active bindings are returned.",
              "title": "Include Deactivated",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Bad request (envelope)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Auth required (envelope)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Forbidden (envelope)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Not found (envelope)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Conflict (envelope)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Validation failed (envelope)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Internal error (envelope)."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Info Item",
        "tags": [
          "info-items"
        ]
      }
    },
...
  }
}